Separated toolbar GUI from Map
