+++ geemap/geemap.py
@@ -274,317 +274,8 @@ class Map(ipyleaflet.Map):
         self.tool_output = tool_output
         tool_output.clear_output(wait=True)
 
-        tools = {
-            "info": {"name": "inspector", "tooltip": "Inspector"},
-            "bar-chart": {"name": "plotting", "tooltip": "Plotting"},
-            "eraser": {
-                "name": "eraser",
-                "tooltip": "Remove all drawn features",
-            },
-            "folder-open": {
-                "name": "open_data",
-                "tooltip": "Open local vector/raster data",
-            },
-            "retweet": {
-                "name": "convert_js",
-                "tooltip": "Convert Earth Engine JavaScript to Python",
-            },
-            "gears": {
-                "name": "whitebox",
-                "tooltip": "WhiteboxTools for local geoprocessing",
-            },
-            "google": {
-                "name": "geetoolbox",
-                "tooltip": "GEE Toolbox for cloud computing",
-            },
-            "map": {
-                "name": "basemap",
-                "tooltip": "Change basemap",
-            },
-            "globe": {
-                "name": "timelapse",
-                "tooltip": "Create timelapse",
-            },
-            "fast-forward": {
-                "name": "timeslider",
-                "tooltip": "Activate timeslider",
-            },
-            "hand-o-up": {
-                "name": "draw",
-                "tooltip": "Collect training samples",
-            },
-            "line-chart": {
-                "name": "transect",
-                "tooltip": "Creating and plotting transects",
-            },
-            "random": {
-                "name": "sankee",
-                "tooltip": "Sankey plots",
-            },
-            "adjust": {
-                "name": "planet",
-                "tooltip": "Planet imagery",
-            },
-            "info-circle": {
-                "name": "cog-inspector",
-                "tooltip": "Get COG/STAC pixel value",
-            },
-            # "spinner": {
-            #     "name": "placehold2",
-            #     "tooltip": "This is a placehold",
-            # },
-            # "question": {
-            #     "name": "help",
-            #     "tooltip": "Get help",
-            # },
-        }
-
-        icons = list(tools.keys())
-        tooltips = [item["tooltip"] for item in list(tools.values())]
-
-        icon_width = "32px"
-        icon_height = "32px"
-        n_cols = 3
-        n_rows = -int(-(len(icons) / n_cols))
-
-        toolbar_grid = widgets.GridBox(
-            children=[
-                widgets.ToggleButton(
-                    layout=widgets.Layout(
-                        width="auto", height="auto", padding="0px 0px 0px 4px"
-                    ),
-                    button_style="primary",
-                    icon=icons[i],
-                    tooltip=tooltips[i],
-                )
-                for i in range(len(icons))
-            ],
-            layout=widgets.Layout(
-                width="109px",
-                grid_template_columns=(icon_width + " ") * n_cols,
-                grid_template_rows=(icon_height + " ") * n_rows,
-                grid_gap="1px 1px",
-                padding="5px",
-            ),
-        )
-        self.toolbar = toolbar_grid
-
-        def tool_callback(change):
-            if change["new"]:
-                current_tool = change["owner"]
-                for tool in toolbar_grid.children:
-                    if tool is not current_tool:
-                        tool.value = False
-                tool = change["owner"]
-                tool_name = tools[tool.icon]["name"]
-                if tool_name == "eraser":
-                    self.remove_drawn_features()
-                    tool.value = False
-                elif tool_name == "inspector":
-                    if not hasattr(self, "inspector_control"):
-                        self.add_inspector()
-                    self.toolbar_reset()
-                elif tool_name == "plotting":
-                    self.plot_checked = True
-                    plot_dropdown_widget = widgets.Dropdown(
-                        options=list(self.ee_raster_layer_names),
-                    )
-                    plot_dropdown_widget.layout.width = "18ex"
-                    self.plot_dropdown_widget = plot_dropdown_widget
-                    plot_dropdown_control = ipyleaflet.WidgetControl(
-                        widget=plot_dropdown_widget, position="topright"
-                    )
-                    self.plot_dropdown_control = plot_dropdown_control
-                    self.add(plot_dropdown_control)
-                    if self.draw_control in self.controls:
-                        self.remove_control(self.draw_control)
-                    self.add_draw_control_lite()
-                elif tool_name == "open_data":
-                    from .toolbar import open_data_widget
-
-                    open_data_widget(self)
-                elif tool_name == "convert_js":
-                    from .toolbar import convert_js2py
-
-                    convert_js2py(self)
-                elif tool_name == "whitebox":
-                    import whiteboxgui.whiteboxgui as wbt
-
-                    tools_dict = wbt.get_wbt_dict()
-                    wbt_toolbox = wbt.build_toolbox(
-                        tools_dict,
-                        max_width="800px",
-                        max_height="500px",
-                        sandbox_path=self.sandbox_path,
-                    )
-                    wbt_control = ipyleaflet.WidgetControl(
-                        widget=wbt_toolbox, position="bottomright"
-                    )
-                    self.whitebox = wbt_control
-                    self.add(wbt_control)
-                elif tool_name == "geetoolbox":
-                    from .toolbar import build_toolbox, get_tools_dict
-
-                    tools_dict = get_tools_dict()
-                    gee_toolbox = build_toolbox(
-                        tools_dict, max_width="800px", max_height="500px"
-                    )
-                    geetoolbox_control = ipyleaflet.WidgetControl(
-                        widget=gee_toolbox, position="bottomright"
-                    )
-                    self.geetoolbox = geetoolbox_control
-                    self.add(geetoolbox_control)
-
-                elif tool_name == "basemap":
-                    from .toolbar import change_basemap
-
-                    change_basemap(self)
-                elif tool_name == "timelapse":
-                    from .toolbar import timelapse_gui
-
-                    timelapse_gui(self)
-                    self.toolbar_reset()
-                elif tool_name == "timeslider":
-                    from .toolbar import time_slider
-
-                    time_slider(self)
-                    self.toolbar_reset()
-                elif tool_name == "draw":
-                    from .toolbar import collect_samples
-
-                    self.training_ctrl = None
-                    collect_samples(self)
-                elif tool_name == "transect":
-                    from .toolbar import plot_transect
-
-                    plot_transect(self)
-                elif tool_name == "sankee":
-                    from .toolbar import sankee_gui
-
-                    sankee_gui(self)
-                elif tool_name == "planet":
-                    from .toolbar import split_basemaps
-
-                    split_basemaps(self, layers_dict=planet_tiles())
-                    self.toolbar_reset()
-                elif tool_name == "cog-inspector":
-                    from .toolbar import inspector_gui
-
-                    inspector_gui(self)
-
-                elif tool_name == "help":
-                    import webbrowser
-
-                    webbrowser.open_new_tab("https://geemap.org")
-                    current_tool.value = False
-            else:
-                tool = change["owner"]
-                tool_name = tools[tool.icon]["name"]
-                if tool_name == "inspector":
-                    pass
-                elif tool_name == "plotting":
-                    self.plot_checked = False
-                    plot_dropdown_widget = self.plot_dropdown_widget
-                    plot_dropdown_control = self.plot_dropdown_control
-                    if plot_dropdown_control in self.controls:
-                        self.remove_control(plot_dropdown_control)
-                    del plot_dropdown_widget
-                    del plot_dropdown_control
-                    if self.plot_control in self.controls:
-                        plot_control = self.plot_control
-                        plot_widget = self.plot_widget
-                        self.remove_control(plot_control)
-                        self.plot_control = None
-                        self.plot_widget = None
-                        del plot_control
-                        del plot_widget
-                    if (
-                        self.plot_marker_cluster is not None
-                        and self.plot_marker_cluster in self.layers
-                    ):
-                        self.remove_layer(self.plot_marker_cluster)
-                    if self.draw_control_lite in self.controls:
-                        self.remove_control(self.draw_control_lite)
-                    self.add(self.draw_control)
-                elif tool_name == "whitebox":
-                    if self.whitebox is not None and self.whitebox in self.controls:
-                        self.remove_control(self.whitebox)
-                elif tool_name == "convert_js":
-                    if (
-                        self.convert_ctrl is not None
-                        and self.convert_ctrl in self.controls
-                    ):
-                        self.remove_control(self.convert_ctrl)
-
-        for tool in toolbar_grid.children:
-            tool.observe(tool_callback, "value")
-
-        toolbar_button = widgets.ToggleButton(
-            value=False,
-            tooltip="Toolbar",
-            icon="wrench",
-            layout=widgets.Layout(
-                width="28px", height="28px", padding="0px 0px 0px 4px"
-            ),
-        )
-        self.toolbar_button = toolbar_button
-
-        layers_button = widgets.ToggleButton(
-            value=False,
-            tooltip="Layers",
-            icon="server",
-            layout=widgets.Layout(height="28px", width="72px"),
-        )
-
-        toolbar_widget = widgets.VBox()
-        toolbar_widget.children = [toolbar_button]
-        toolbar_header = widgets.HBox()
-        toolbar_header.children = [layers_button, toolbar_button]
-        toolbar_footer = widgets.VBox()
-        toolbar_footer.children = [toolbar_grid]
-
-        toolbar_event = ipyevents.Event(
-            source=toolbar_widget, watched_events=["mouseenter", "mouseleave"]
-        )
-
-        def handle_toolbar_event(event):
-            if event["type"] == "mouseenter":
-                toolbar_widget.children = [toolbar_header, toolbar_footer]
-            elif event["type"] == "mouseleave":
-                if not toolbar_button.value:
-                    toolbar_widget.children = [toolbar_button]
-                    toolbar_button.value = False
-                    layers_button.value = False
-
-        toolbar_event.on_dom_event(handle_toolbar_event)
-
-        def toolbar_btn_click(change):
-            if change["new"]:
-                layers_button.value = False
-                toolbar_widget.children = [toolbar_header, toolbar_footer]
-            else:
-                if not layers_button.value:
-                    toolbar_widget.children = [toolbar_button]
-
-        toolbar_button.observe(toolbar_btn_click, "value")
-
-        def layers_btn_click(change):
-            if change["new"]:
-                from .toolbar import layer_manager_gui
-
-                # Create Layer Manager Widget
-                toolbar_footer.children = layer_manager_gui(self, return_widget=True)
-            else:
-                toolbar_footer.children = [toolbar_grid]
-
-        layers_button.observe(layers_btn_click, "value")
-        toolbar_control = ipyleaflet.WidgetControl(
-            widget=toolbar_widget, position="topright"
-        )
-
         if kwargs.get("toolbar_ctrl"):
-            self.add(toolbar_control)
-            self.toolbar_ctrl = toolbar_control
+            self.add_toolbar()
 
         def handle_interaction(**kwargs):
             latlon = kwargs.get("coordinates")
@@ -6988,6 +6679,11 @@ class Map(ipyleaflet.Map):
             remove=False,
         )
         self.draw_control_lite = draw_control_lite
+
+    def add_toolbar(self, position="topright", **kwargs):
+        from .toolbar import main_toolbar
+
+        main_toolbar(self, position, **kwargs)
 
 
 # The functions below are outside the Map class.
